Added Q1 .jar
Note: This jar currently works on a *static DataFrame*, so twitter data will need to exist in the twitterstream folder for anything to happen. It will download data on its own as it runs, but there must be something in the folder on the first run for it to work.