Added Q3 .jar
Note: This is a .jar file of the code as it is on GitHub, meaning, it REQUIRES a twitterstream folder, and data IN that twitterstream folder to work.  
To stream new data, you will need to uncomment the blocks for the future methods, and re-assemble the file (sbt assembly)  
If you wish to see live updates as the data is streamed in, replace the staticDF block with the commented, streamDF block.